scrapers/requirements.txt
psycopg2-binary>=2.9.9
pydantic>=2.0.0
python-dotenv>=1.0.0
playwright>=1.58.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
